Moving utils into new folder with separate files
