+++ app/src/organisms/QuickTransferFlow/VolumeEntry.tsx
@@ -10,7 +10,7 @@ import {
 import { ChildNavigation } from '../ChildNavigation'
 import { InputField } from '../../atoms/InputField'
 import { NumericalKeyboard } from '../../atoms/SoftwareKeyboard'
-import { getVolumeLimits } from './utils'
+import { getVolumeRange } from './utils'
 import { CONSOLIDATE, DISTRIBUTE } from './constants'
 
 import type { SmallButton } from '../../atoms/buttons'
@@ -35,7 +35,7 @@ export function VolumeEntry(props: VolumeEntryProps): JSX.Element {
   const [volume, setVolume] = React.useState<string>(
     state.volume ? state.volume.toString() : ''
   )
-  const volumeRange = getVolumeLimits(state)
+  const volumeRange = getVolumeRange(state)
   let headerCopy = t('set_transfer_volume')
   let textEntryCopy = t('volume_per_well_µL')
   if (state.transferType === CONSOLIDATE) {
